Responding to feedback: moving deploy.go -> minikubetestenv, and improving new client helper functions
