+++ src/internal/minikubetestenv/deploy.go
@@ -1,4 +1,4 @@
-package testutil
+package minikubetestenv
 
 import (
 	"context"
@@ -12,6 +12,7 @@ import (
 	"github.com/pachyderm/pachyderm/v2/src/internal/backoff"
 	"github.com/pachyderm/pachyderm/v2/src/internal/errors"
 	"github.com/pachyderm/pachyderm/v2/src/internal/require"
+	"github.com/pachyderm/pachyderm/v2/src/internal/testutil"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	kube "k8s.io/client-go/kubernetes"
@@ -68,13 +69,13 @@ func waitForPachd(t *testing.T, ctx context.Context, kubeClient *kube.Clientset,
 func InstallPublishedRelease(t *testing.T, ctx context.Context, kubeClient *kube.Clientset, version, user string) *client.APIClient {
 	require.NoError(t, helm.InstallE(t, localDeploymentWithMinioOptions(ns, version), helmChartPublishedPath, helmRelease))
 	waitForPachd(t, ctx, kubeClient, ns, version)
-	return GetNewAuthenticatedPachClient(t, user)
+	return testutil.AuthenticatedPachClient(t, testutil.NewPachClient(t), user)
 }
 
 func UpgradeRelease(t *testing.T, ctx context.Context, kubeClient *kube.Clientset, user string) *client.APIClient {
 	require.NoError(t, helm.UpgradeE(t, localDeploymentWithMinioOptions(ns, localImage), helmChartLocalPath, helmRelease))
 	waitForPachd(t, ctx, kubeClient, ns, localImage)
-	return GetNewAuthenticatedPachClient(t, user)
+	return testutil.AuthenticatedPachClient(t, testutil.NewPachClient(t), user)
 }
 
 func DeleteRelease(t *testing.T, ctx context.Context, kubeClient *kube.Clientset) {
